Updated the UI
I made changes to structure of the UI module, as well as adding a hitbox
module.  The hitbox module is intended to have several hitbox classes
for use with the UI.

The UI module was updated to include a base class for the common UI
elements.  As the two methods of using the UI both have many
commonalities, using a base class will help reduce bugs.  The only
difference between the two methods is how either method responds to
pygame events.  One uses a callback function, while the other uses
inheritance to override a select few methods for handling various
events.